pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gitdocs"
version = "0.1.0"
description = "Developer-friendly CLI/TUI for Jira tickets and Confluence documentation"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "gitdocs team" }
]
keywords = ["jira", "confluence", "cli", "tui", "developer-tools", "documentation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Documentation",
]

dependencies = [
    "typer[all]>=0.9.0",
    "textual>=0.47.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.26.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
    "keyring>=24.3.0",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.3",
    "aiofiles>=23.2.1",
    "tenacity>=8.2.3",
    "diskcache>=5.6.3",
    "markdownify>=0.11.6",
    "markdown>=3.5.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "respx>=0.20.2",
    "ruff>=0.1.13",
    "mypy>=1.8.0",
    "types-PyYAML>=6.0.12",
]
llm = [
    "openai>=1.9.0",
    "tiktoken>=0.5.2",
]

[project.scripts]
gitdocs = "gitdocs.cli.app:app"

[project.urls]
Homepage = "https://github.com/gitdocs/gitdocs"
Documentation = "https://gitdocs.dev"
Repository = "https://github.com/gitdocs/gitdocs"

[tool.hatch.build.targets.wheel]
packages = ["gitdocs"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

